Sbs-bot-Fuck-Json-Files-update-main/Sbs-bot-Fuck-Json-Files-update-main/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0   
aiohttp>=3.8.5
aiofiles>=23.1.0
pydub>=0.25.1
supabase>=2.22.0
aiosqlite>=0.17.0
requests>=2.32.5
